backend/requirements.txt
Flask==2.2.3
requests==2.28.2
openai==0.27.0
spotipy==2.22.1
transformers==4.27.4
pandas==1.5.3
Flask-Cors==5.0.1
python-dotenv==0.19.2
